+++ lib/router/list_routes.dart
@@ -1,6 +1,7 @@
 import 'package:go_router/go_router.dart';
 
 import '../ui/layouts/main_layout/main_layout_page.dart';
+import '../ui/views/home_view/home_view.dart';
 import 'path_routes.dart';
 
 final listRoutes = [
@@ -10,7 +11,9 @@ final listRoutes = [
     pageBuilder: (context, state) {
       // key: state.pageKey;
       return const NoTransitionPage(
-        child: MainLayoutPage()
+        child: MainLayoutPage(
+          child: HomeView(),
+        )
       );
     },
   ),
