+++ lib/ui/layouts/main_layout/main_layout_page.dart
@@ -1,12 +1,15 @@
 import 'package:flutter/material.dart';
-import '../../shared/custom_app_menu.dart';
-import 'views/about_view.dart';
-import 'views/hero_view.dart';
-import 'views/services_view.dart';
-import 'views/why_view.dart';
+
+import '../../shared/custom_app_menu/custom_app_menu.dart';
+
 
 class MainLayoutPage extends StatelessWidget {
-  const MainLayoutPage({super.key});
+  const MainLayoutPage({
+    super.key,
+    required this.child,
+  });
+
+  final Widget child; 
 
   @override
   Widget build(BuildContext context) {
@@ -14,37 +17,17 @@ class MainLayoutPage extends StatelessWidget {
       body: Container(
         child: Stack(
           children: [
-            _MainLayoutPage(),
+            child,
             Positioned(
-              
               child: Container(
                 width: MediaQuery.of(context).size.width,
                 color: Colors.transparent,
-                child: CustomAppMenu(),
-              ) 
+                child: const CustomAppMenu(),
+              ),
             )
           ],
         ),
       ),
-    );
-  }
-}
-
-
-class _MainLayoutPage extends StatelessWidget {
-  const _MainLayoutPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return ListView(
-      scrollDirection: Axis.vertical,
-      //
-      children: [
-        HeroView(),
-        ServicesView(),
-        WhyView(),
-        AboutView()
-      ],
     );
   }
 }
