+++ lib/ui/views/home_view/sections/about_section.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class AboutView extends StatelessWidget {
-  const AboutView({super.key});
+class AboutSection extends StatelessWidget {
+  const AboutSection({super.key});
 
   @override
   Widget build(BuildContext context) {
