+++ lib/ui/views/home_view/sections/utils/gradient_text.dart
